Added wx support for examples/tasks/advanced example * changed to EditorAreaPane as SplitEditorAreaPane isn't supported in wx yet * added runtime selection of toolkit in python_editor.py
Conflicts:
	examples/tasks/advanced/python_editor.py
